requirements.txt
acme==1.21.0
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
arrow==1.3.0
Authlib==1.6.0
bcrypt==3.2.0
binaryornot==0.4.4
blinker==1.4
Brlapi==0.8.3
cachetools==5.5.2
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
cliapp==1.20180812.1
click==8.2.1
cloudpickle==3.1.1
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
cookiecutter==2.6.0
cryptography==45.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docstring_parser==0.16
duplicity==0.8.21
exceptiongroup==1.2.2
fastapi==0.115.13
fasteners==0.14.1
flet==0.26.0
flet-cli==0.26.0
flet-desktop==0.26.0
future==0.18.2
google-adk==1.4.1
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.21.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
jeepney==0.7.1
Jinja2==3.1.5
josepy==1.10.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mcp==1.9.4
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==24.2
pandas==2.0.1
paramiko==2.9.3
parsedatetime==2.6
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.8.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.3
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.25.3
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.30.1
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
reportlab==3.6.8
requests==2.32.4
requests-toolbelt==0.9.1
rich==13.9.4
rsa==4.9.1
SecretStorage==3.3.1
shapely==2.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
systemd-python==234
tenacity==8.5.0
text-unidecode==1.3
toml==0.10.2
ttystatus==0.38
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.3.0
usb-creator==0.3.7
uv==0.7.13
uvicorn==0.34.3
wadllib==1.3.6
watchdog==4.0.2
websockets==15.0.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==3.23.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
